entryManager:CPUの追加処理の作成（見た目はカス） middleResultManager:勝利ポイントの桁数が謎のfloatになってる問題解決 player：Createの引数にAIの設定を追加 retentionManager:AIのON/OFF判定追加 scene:AI状況の設定を生成時に設定 sceneEntry:AI状況の設定を生成時に設定

+++ 00_ソースファイル/project000_LAMP/middleResultManager.cpp
@@ -66,7 +66,7 @@ namespace
 		const D3DXVECTOR3 SIZE_TITLE	= D3DXVECTOR3(300.0f, 100.0f, 0.0f);	// タイトル大きさ
 		const D3DXVECTOR3 SIZE_VALUE	= D3DXVECTOR3(100.0f, 100.0f, 0.0f);	// 数字大きさ
 
-		const float	DIGIT = 2;	// 桁数
+		const int	DIGIT = 2;	// 桁数
 	}
 
 	// 勝利ポイント背景基本情報
